isl_pw_multi_aff_from_map: avoid computing all equalities if not needed
If the available equality constraints are enough to determine that
the input map is obviously single valued (in which case the isl_pw_multi_aff
can be easily derived from the input), then there is no need to look
for any further equality constraints.  In particular, there is no need
to find all valid equality constraints.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>