isl_map_simplify.c: drop_more_redundant_divs: no need to check unit constraints
The function drop_more_redundant_divs looks for existentially quantified
variables that can be eliminated without introducing any spurious solutions.
In particular, for each pair of lower and upper bounds on the variable,
it checks that there is at least one integer value between the two bounds.
If at least one of the two constraint has a unit coefficient for the variable,
then the presence of such a value is guaranteed so there is no need to check.
Note that this is a special case of the Omega test.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>