isl_map_simplify.c: int_between_bounds: create tableau on demand
When looking for redundant existentially quantified variables
by checking that each pair of lower and upper bounds admits an intermediate
integer value, there may be no need to evaluate any test inequality
explicitly.  This happens in particular when the Omega test applies.
If so, there is no need to create a tableau in the first place.
Only create it when needed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>