isl_printer_print_map: print div constraints again in Omega format
Since f164818 (isl_printer_print_map: do not print div constraints,
Wed Aug 26 15:35:11 2015 +0200) div constraints are no longer printed
because they are redundant with respect to the explicit definition
of the integer division.  However, in Omega format, the explicit
definition is never printed, so the div constraints are not redundant
in that case and should still be printed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>