isl_coalesce.c: generalize check_coalesce_subset to coalesce_divs
That is, do not only check for one basic map being a subset of the other,
but also allow the integer divisions to be exploited in other ways.
This commit does not change the behavior of the function, but only
makes it possible for it to be extended to other cases in later commits.
This means, in particular, that the subfunction should not just
indicate whether one basic map was determined to be a subset of the other,
but instead should allow for any isl_change.  For consistency, this also
means that the basic map that is determined to be a subset of the other
is dropped immediately in contains_with_expanded_divs (now called
coalesce_with_expanded_divs).
This in turn means that the full isl_coalesce_info data structure
needs to be passed down to this function.
The basic map is still passed in separately since it may have
been modified prior to the call to coalesce_after_aligning_divs.
In particular, this happens inside coalesced_subset_with_equalities
(now called coalesce_subset_with_equalities).

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>