isl_coalesce.c: coalesce_local_pair: allow reuse of status information
In a future commit, coalesce_local_pair will be called from a context
where the status of the constraints of one basic map with respect
to the other basic map has already been computed.  Allow this
information to be reused by not recomputing it when the corresponding
info fields have already been set.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>