isl_coalesce.c: wrap_in_facets: do not pass in list of cut constraints
The set of cut constraints is already available in info[i].ineq,
so there is no need to pass it in separately.
In an upcoming commit, a basic map will also be allowed to protrude
out of half of an equality constraint.  Explicitly passing in
the list of such cut constraints as well, would complicate the code.
Just remove the explicit list of cut constraints instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>