isl_map_coalesce: allow protrusions out of equality constraints
The original code would only allow a basic map to stick out of
the inequality constraints of another basic map by one.
Also allow it to stick out of the equality constraints by one.

For example, the set { [x, -x] : 0 <= x <= 1 } sticks out
of one side of the equality constraint x = 1 of
{ [1, y] : -1 <= y <= 1 } and can be wrapped in.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>