isl_map_coalesce: check for protrusions on adjacent equality
The set { [x, -x] : 0 <= x <= 1 } has an equality that is adjacent
to an inequality of { [1, y] : 0 <= y <= 1 }, but that inequality
is not considered to be adjacent to an equality of the first set.
coalesce_local_pair_reuse decides to check for the pattern of
adjacent equality and inequality based on the equality, but
the checks in check_adj_eq fail, so no coalescing is performed.
However, some other patterns may still apply.  In particular,
the set with the equality may still be a protrusion out of
the other set, as in this example.
Check for this pattern in check_adj_eq when no corresponding
inequality can be found.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>