isl_coalesce.c: add_sub_vars: explicitly mark added divs as unknown
The divs were neither given an explicit definition, nor marked as unknown,
meaning that they could get some random explicit definition.
This probably does not cause any real problems, because fuse only
copies definitions if they are equal in both basic maps.
However, if this random explicit definition just happens to be equal
to that in the other basic map, then correctness is not guaranteed
since the div constraints to not appear in this basic map and they
may have been simplified away in the other basic map with respect
to another constraint that is later removed.

Also copy some information from the commit message of 172b680
(isl_coalesce.c: coalesce_with_subs: do not add div constraints
to basic map, Wed Jul 15 11:01:01 2015 +0200) to the documentation
to clarify why the div constraints are not added.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>