isl_coalesce.c: fuse: call isl_basic_map_finalize
Do so instead of manually marking the basic map as final.
It is not clear why the original implementation in c8559a0
(add isl_set_coalesce, Mon Mar 16 11:58:57 2009 +0100) resorted
to this direct manipulation.  The only difference between
a call to isl_basic_map_finalize (or isl_basic_set_finalize) and
setting the flag directly is that isl_basic_map_finalize
also removes redundant integer divisions.  Removing such
redundant integer divisions should not cause any harm,
because the tableau is recomputed anyway.  Instead,
it should make any subsequent computation a bit easier.
In contrived cases, as in the new test case, it could even result
in extra subsequent coalescings that would not be possible
in the presence of the redundant integer divisions.
Perhaps it was considered unlikely for an integer division
to become redundant after fusing, but it is certainly possible.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>