isl_tab_basic_map_partial_lexopt: limit scope of symmetry detection
In particular, only perform symmetry detection on input variables
that only occur once.

a41dca9 (isl_tab_basic_map_partial_lexopt: detect and exploit simple
symmetries, Mon Dec 13 16:45:44 2010 +0100) introduced a simple
mechanism for detecting and exploiting symmetries that works
fairly well for inputs such as testsets/pip/cnt_sum2.pip and
testsets/pip/jcomplex.pip.  However, it results in
a significant slow-down on inputs such as

    { S_2[i, k, j] -> [o0, o1, o2, o3] : exists (e0 = floor((o1)/8):
    	8e0 = o1 and k <= 255 and o0 <= 255 and o2 <= 255 and
    	o3 <= 255 - j and 7o3 >= 255 - j and 7o3 <= 7 - i and
    	240o3 <= 239 + o0 and 247o3 <= 247 + k - j and
    	247o3 <= 247 + k - o1 and 247o3 <= 247 + i and
    	248o3 >= 248 - o1 and 248o3 <= o2 and 254o3 >= i - o0 + o1
    	and 254o3 >= -o0 + o1 and 255o3 >= -i + o0 - o1 and
    	1792o3 >= -63736 + 257o1) }

which was reported by Wenlei Bao and which has been added as a new
test case.

The symmetry detection looks for sequences of constraints
with the same coefficients for the output variables,
i.e., those for which the minimum is being computed, and
replaces the sequence by a single constraint with as bound
a new parameter that abstracts the set of original bounds.
If the input variables that appear in the original bounds
also occur in other constraints, then the connection between
those constraints and the modified constraint is lost
since the expression containing the input variable is replaced
by a new parameter in the modified constraint.
This may result in more time being spent on exploring
the extended context.

Only apply the symmetry detection to constraints that only
involve input variables that only occur in those constraints.
This prevents the potential problematic behavior described
above.  It may perhaps also remove some useful applications
of the symmetry detection.  The heuristic may therefore
still have to be refined further based on other inputs.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>