isl_tab_pip.c: propagate_equalities: remove undo tracking when done
propagate_equalities temporarily needs undo tracking in the main tableau
to test whether an equality can be propagated to this main tableau.
The undo tracking should however be turned off again when
propagate_equalities has finished its computation as otherwise
redundant constraints may get needlessly piled up in the main tableau.
Note that undo tracking is not otherwise needed in the main tableau
because a copy of the tableau is created when a choice is made.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>