isl_basic_map_lexopt*: postpone extraction of domain
When isl_basic_map_lexopt* is used to compute a full optimum,
a domain is first extracted from the input and then those constraints
are added back to the input in isl_tab_basic_map_partial_lexopt*.
Avoid this needless intersection by postponing the extraction
of the domain until isl_tab_basic_map_partial_lexopt* such that it can
tell if the intersection is needed.

The intersection that is avoided in this way should be fairly
harmless since the constraints that would be added are identical
to some of the constraints that already in the description of the input.
However, in a subsequent commit, the domain will optionally be computed
more accurately and in that case the intersection could introduce
additional (redundant) constraints that may involve additional
existentially quantified variables.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>