isl_*_lexopt: use actual domain of input as initial context
Since a47882c (isl_basic_map_lexopt: preinitialize domain,
Tue Mar 12 17:23:49 2013 +0100), isl_basic_map_lexopt computes
the optimum over a domain derived from the input rather than
simply over the universe domain.  This means that the context
typically starts out with some constraints and fewer splits
of the context are needed to compute a solution.
The computation of this domain has since been generalized
to other variations of isl_basic_map_lexopt and is now computed
by extract_domain in isl_tab_pip.c.

The resulting initial context can be further refined
to the actual domain of the input rather than some overapproximation
of this domain.  Using the actual domain avoids introducing splits
that only serve to separate off some parts of the domain where
there is no solution.  The drawback of using the actual domain
is that the description of this domain may involve existentially
quantified variables, which means that the tableau will have
a greater number of variables.  However, the extra computations
needed in the context should on average be offset by the savings
obtained by the reduction in the number of splits.

The existentially quantified variables may also end up in the result,
but this should not be a problem when computing lexicographic optima
as such since those variables can still be eliminated afterwards.
However, the elimination of existentially quantified variables
in turn relies on parametric integer programming.  This means
that if it is used for this purpose, an approximation that
is guaranteed to be free of existentially quantified variables
should continue to be used.

For example, computing the lexicographic minimum of the input

    { S_2[i, k, j] -> [o0, o1, o2, o3] : exists (e0 = floor((o1)/8):
	    8e0 = o1 and k <= 255 and o0 <= 255 and o2 <= 255 and
	    o3 <= 255 - j and 7o3 >= 255 - j and 7o3 <= 7 - i and
	    240o3 <= 239 + o0 and 247o3 <= 247 + k - j and
	    247o3 <= 247 + k - o1 and 247o3 <= 247 + i and
	    248o3 >= 248 - o1 and 248o3 <= o2 and 254o3 >= i - o0 + o1
	    and 254o3 >= -o0 + o1 and 255o3 >= -i + o0 - o1 and
	    1792o3 >= -63736 + 257o1) }

would result in

    { S_2[i, k, j] -> [o0 = -127762 + i + 502j, o1 = -62992 + 248j, o2 = 63240 - 248j, o3 = 255 - j] : 502j <= 128017 - i and ((0 < i <= 2 and k <= 255 and j <= 84 + k and 742j >= 188724 - i) or (j >= 85 + k and 1779 + i <= 7j < 247i + 7k) or (i >= 3 and k <= 255 and j <= 84 + k and 7j >= 1779 + i) or (i > 0 and 85 + k <= j <= 247 + k and 7j >= 247i + 7k and 246j >= 62739 - k));
	 S_2[i, k, j] -> [o0 = 248, o1 = 248, o2 = 0, o3 = 0] : 3*floor((j + 2*floor((j + 3*floor((8k + 239j)/247))/4))/3) = j + 2*floor((j + 3*floor((8k + 239j)/247))/4) and 3*floor((2j + floor((j + 3*floor((8k + 239j)/247))/4))/3) = 2j + floor((j + 3*floor((8k + 239j)/247))/4) and i <= 0 and j >= 85 + k and j >= 255 and -i + k <= j <= 247 + k and 240j >= 61256 - 7k and 246j <= 62738 - k and 7*floor((j + 3*floor((8k + 239j)/247))/4) >= 1488 + j and 509*floor((j + 3*floor((8k + 239j)/247))/4) >= -3054 - 6i + 509j;
	 S_2[i, k, j] -> [o0 = 248 + i, o1 = 248, o2 = 0, o3 = 0] : i = 7 and j = 255 and floor((2)/1) = 2 and k <= 255 and (k >= 171 or (k <= 170 and 7k >= 57));
	 S_2[i = 0, k, j = 251] -> [o0 = 1, o1 = 0, o2 = 248, o3 = 1] : 251 <= k <= 255;
	 S_2[i, k, j] -> [o0 = -127762 + 502j, o1 = -62992 + 248j, o2 = 63240 - 248j, o3 = 255 - j] : (i = 0 and j = 255 and 171 <= k <= 255) or (i <= 0 and j >= 85 + k and 246j >= 62739 - k and 509j <= 129795 + i);
	 S_2[i = 0, k, j] -> [o0 = 1, o1 = 0, o2 = 248, o3 = 1] : k <= 255 and 248 <= j <= 254 and j <= k and (j <= 250 or j >= 252);
	 S_2[i, k, j] -> [o0 = 248 + i, o1 = 248, o2 = 0, o3 = 0] : 3*floor((j + 2*floor((j + 3*floor((8k + 239j)/247))/4))/3) = j + 2*floor((j + 3*floor((8k + 239j)/247))/4) and 3*floor((2j + floor((j + 3*floor((8k + 239j)/247))/4))/3) = 2j + floor((j + 3*floor((8k + 239j)/247))/4) and i > 0 and k <= 255 and j >= 85 + k and 255 <= j <= 247 + k and 7j >= 247i + 7k and 246j <= 62738 - k and 7*floor((j + 3*floor((8k + 239j)/247))/4) >= 1488 + j }

and now results in

    { S_2[i, k, j] -> [o0 = -127762 + i + 502j, o1 = -62992 + 248j, o2 = 63240 - 248j, o3 = 255 - j] : exists (e0: k <= 255 and 106j >= 26544 - i and 742j >= 188724 - i and e0 <= 255 - j and 255 - j <= 7e0 <= 7 - i and 247e0 <= 247 + k - j and 247e0 <= 247 + i and 509e0 >= -i);
	 S_2[i = 0, k, j] -> [o0 = 1, o1 = 0, o2 = 248, o3 = 1] : k <= 255 and 248 <= j <= 254 and j <= k and (j >= 251 or j <= 250) }

Note that this new result does in fact involve an existentially
quantified variable.  In this particular example, the existentially
quantified variable happens to be equal to 255 - j, such that the result
could be further simplified to

    { S_2[i, k, j] -> [o0 = -127762 + i + 502j, o1 = -62992 + 248j, o2 = 63240 - 248j, o3 = 255 - j] : k <= 255 and 7j >= 1778 + i and 246j >= 62738 - k and 247j >= 62738 - i and 509j <= 129795 + i and 742j >= 188724 - i;
	 S_2[i = 0, k, j] -> [o0 = 1, o1 = 0, o2 = 248, o3 = 1] : k <= 255 and 248 <= j <= 254 and j <= k and (j >= 251 or j <= 250)

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>