isl_*_partial_lex{min,max}_pw_multi_aff: handle existentially quantified vars
If the context involves existentially quantified variables without
a known explicit representation, then the construction of the solution
would fail because the affine expressions are constructed on top
of the local space of the (extended) context and an affine expression
is not allowed to refer to existentially quantified variables without
a known explicit representation.

Move those variables last such that they can be dropped from
the local space without losing the correspondence between
the variables in the local space and the columns in the matrix.
Exploit the equality constraints in the extended context to ensure
that the last columns in the matrix (i.e., those that will
be ignored) are zero.  These always exist as this is
the only way the matrix could refer to those variables
in the first place.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>