isl_coalesce.c: is_adj_eq_extension: check extension along multiple constraints
That is, allow some cut constraints along with the inequality constraint
that is adjacent to an equality.  Change the name to is_relaxed_extension
to reflect the extended functionality.  The next commit will modify the caller
to collect the cut constraints and therefore to exploit this extended
functionality.

Note that the configuration depicted in the new ASCII art is
unlikely to happen in practice because the vertical cut constraint
would have have been removed from the basic map that is about
to be extended because it is redundant with respect to the two
slanted constraints.  However, cut constraints can still end up
being extended in similar configurations, as illustrated by some
of the new test cases added in the next commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>