add isl_tab_is_constant
This function checks whether any given integer variable can only attain
a single integer value, adding the appropriate equality constraint.
It can be seen as a cheaper form of isl_tab_detect_equalities
that is specialized to only check if the single variable is equal to
a fixed integer value rather than an affine expression.

This function will be used in the next commit to check
whether integer divisions that are copied from one basic map
to another have fixed values in this other basic map.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>