isl_map_coalesce: detect constants among expanded divs
Since a130b58 (isl_map_coalesce: allow general coalescing with expanded divs,
Fri May 27 13:12:48 2016 +0200), coalescing opportunities are checked
in pairs of basic maps where one of them has extra integer divisions
by first copying these to the other basic map.
These extra integer divisions may in some cases only attain
a single integer value.  Detect those cases and add corresponding
equality constraints to the tableau.  Constraints of the first basic
map that may appear to cut off some values of the second basic map
based on the range of possible rational values can in some cases
be turned into valid constraints in this way since some rational
values are cut off beforehand.

For example, in the new test case

    { [a, b] : a <= 1024 and b >= 0 and
	    ((-31 - a + b <= 32*floor((-1 - a)/32) <= -33 + b and
		32*floor((-1 - a)/32) <= -16 + b + 16*floor((-1 - a)/16)) or
	    (2 <= a <= 15 and b < a)) }

the integer divisions floor((-1 - a)/32) and floor((-1 - a)/16)
both have the fixed value -1.

Care needs to be taken however not to add both the div constraints
and the equality constraints as that may render some of the original
constraints redundant.  These would then be ignored during
the coalescing detection, but may remain in the result.

For example, in the new test case

    { [a] : a > 0 and
	    ((16*floor((a)/16) < a and 32*floor((a)/32) < a) or a <= 15) }

alpha = floor((a)/16) has the value 0 in the second disjunct.
Adding both the div constraints and the constraints alpha >= 0 and alpha <= 0
would allow 15 - a >= 0 to be written as
16 * (-alpha) + 1 * (16 alpha + 15 - a) >= 0.
The constraint a <= 15 would then be treated as redundant, in which
case the first disjunct is a trivial subset of the second and
the final result would be

    { [a] : 0 < a <= 15 }

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>