Add new isl_ast_expr_to_C_str function
This function provides a new interface to return a C code representation for
an isl_ast_expr. It can be used as replacement for isl_ast_expr_to_str in
situations where it is important that the returned representation of an
ast_expr uses C formatting.

This new functionality is directly used in isl_test to make it
future-proof for situations where the isl_ast_expr representation
changes.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>