isl_local_space_get_div: handle presence of unknown local variables
The original implementation would only check that the integer
division has a (partial) explicit representation.  However, it
would not check if the explicit representation is complete and
it would fail if any of the other local variables do not have
an explicit representation.

Properly check that the explicit representation is complete and
drop local variables without an explicit representation such
that the result can be constructed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>